Refs [Task] [9294, 9295] trừ stock của sản phầm,giao diện (zalo pay,áp dụng voucher...)

+++ src/main/java/com/ra/model/entity/OrderDetail.java
@@ -14,7 +14,7 @@ import lombok.*;
 public class OrderDetail {
     @EmbeddedId
     private OrderDetailId id;
-    private int orderQuantity;
+    private Long orderQuantity;
     private String orderDetailName;
     private Double unitPrice;
